fix: resolve test failures and CI issues
- Extract CrewGame from server/index.ts to src/game/CrewGame.ts
- Update all test imports to use extracted game definition
- Fix server import side effects causing port conflicts
- Resolve "Cannot read properties of undefined (reading 'players')" error
- Fix linting errors in CrewGame.ts
